enh(ui): add the bounding box in mesh data
The axis-aligned bounding box of a mesh is now stored directly in `sight::data::mesh`. It is lazily computed when `get_bounding_box()` is called.

It can be used for intersection tests. For instance, a new function `sight::geometry::data::is_inside_bounding_box` was implemented to determine if a 3D point lies inside a bounding box.