feat(core): introduce doctest for unit tests
We will migrate over time from the old CppUnit library to the more modern Doctest.

A new target type DOCTEST was added. It supports the same arguments as CppUnit tests, like REQUIRE_X, ADMIN, etc...

```cmake
sight_add_target(lib_ut TYPE DOCTEST)
```
The tests discovery has been partially rewritten from the original doctest cmake scripts to adapt to our structure, to get the tests available in VSCode testing panel, and to handle child repositories.
For debugging in VSCode, it was impossible to use the previous approach using a launcher script, that already did not work with CppUnit. To overcome this, we no longer use a script and instead pass a modified environment directly to CTest.