refactor(core): new progress API, removed old DICOM code
The former ``sight::core::jobs`` was heavily refactored as ``sight::core::progress``. It is mainly a clarification and simplification of its purpose inside the library _Sight_.

It was previously designed to report the progress **and** to manage the execution of threaded tasks. This second goal was redundant in modern C++ (``std::thread``, ``std::launch``, OpenMP, etc...), made the API cumbersome and often confused the developers, leading to inappropriate usage, especially in services.

The new API no longer comprises task execution. It remains thread-safe but only allows to report progress and check cancellation. Thus, the former sight::core::jobs::job has no equivalent in the new API. We only kept the observer and the aggregator. Here is an example:

```cpp
void progress_service::updating()
{
    auto monitor = std::make_shared<sight::core::progress::monitor>( "Example of Monitor");

    // Emit the signal to inform the progress bar about the new monitor
    this->async_emit(has_monitors::signals::JOB_CREATED, monitor->get_sptr());

    ...
    /// This will be caught by the progress bar to update the widget
    monitor->done_work(i);
    ...

    /// This is a convenient way to set the amount of done work units to total work units (100%)
    monitor->done();
}
```

Also, the API also automatically affects a service using monitors, i.e. which inherits from `sight::core::progress::has_monitors`, to a dedicated worker. This way, by design, the service is asynchronous and does not let a chance to the developer to block the user interface and thus progress report. If a worker is specified, it is used however.

The signals/slots were renamed as well:

```xml
<service uid="progress_srv" type="sight::module::sample::progress_service" />

<service uid="progress_bar_srv" type="sight::module::ui::qt::progress_bar" />

<connect>
    <signal>progress_srv/monitor_created</signal>
    <slot>progress_bar_srv/add_monitor</slot>
</connect>
```

All services reporting progress were ported to the new API, and thus inherits from `sight::core::progress::has_monitors`. The `sight::io::service::reader` and `sight::io::service::writer` also inherit from this interface to encourage developers to report progress. The readers/writers no longer instantiate the "job" but their `read/write` method now take an observer by default. This might be a little too constraining, but I wanted to enforce developers to report progress.

A lot of the old job API was present in the legacy DICOM code. Instead of porting to the new API, the legacy DICOM code was removed. This only part where it was still used was the DIMSE/DicomWeb features, this code was ported to use the modern DICOM code.