enh(ui): add switch button widget
The classic checkbox looked dated, so we introduced a modern toggle switch: it displays the on/off state much more clearly yet behaves exactly like the checkbox behind the scenes.

To use this new widget, in the configuration of a `sight::module::ui::qt::settings` service, specify `widget="switch_button"` for a switch button or `widget="check_box"` for a checkbox when wrapping a `sight::data::boolean`.